refactor(email): simplify preview fixtures
