+++ components/emails/company-invite.tsx
@@ -1,16 +1,13 @@
 import enMessages from "@/i18n/locales/en.json";
 import { EmailButton } from "@/components/emails/base/email-button";
-import { EmailLayout } from "@/components/emails/base/email-layout";
-import { PREVIEW_EMAIL_LAYOUT_COPY, type EmailLayoutCopy } from "@/components/emails/base/email-layout-copy";
+import { EmailLayout, type EmailLayoutSharedProps } from "@/components/emails/base/email-layout";
 import { EmailLink } from "@/components/emails/base/email-link";
 import { EmailSection } from "@/components/emails/base/email-section";
 import { EmailText } from "@/components/emails/base/email-text";
-import { DEFAULT_LOCALE, type AppLocale } from "@/i18n/locale-registry";
+import { PREVIEW_EMAIL_LAYOUT_PROPS } from "@/components/emails/preview-layout-props";
 import { env } from "@/env";
 
-type Props = {
-  locale: AppLocale;
-  layoutCopy: EmailLayoutCopy;
+type Props = EmailLayoutSharedProps & {
   inviteLink: string;
   subject: string;
   preview: string;
@@ -19,18 +16,9 @@ type Props = {
   fallback: string;
 };
 
-export default function CompanyInvite({
-  locale,
-  layoutCopy,
-  inviteLink,
-  subject,
-  preview,
-  intro,
-  cta,
-  fallback,
-}: Props) {
+export default function CompanyInvite({ inviteLink, subject, preview, intro, cta, fallback, ...layoutProps }: Props) {
   return (
-    <EmailLayout layoutCopy={layoutCopy} locale={locale} preview={preview} title={subject}>
+    <EmailLayout {...layoutProps} preview={preview} title={subject}>
       <EmailText>{intro}</EmailText>
 
       <EmailSection>
@@ -50,12 +38,11 @@ const t = enMessages.CompanyInvite;
 const previewInviterName = "Anna Müller";
 
 CompanyInvite.PreviewProps = {
-  layoutCopy: PREVIEW_EMAIL_LAYOUT_COPY,
-  locale: DEFAULT_LOCALE,
+  ...PREVIEW_EMAIL_LAYOUT_PROPS,
   inviteLink: `${env.BASE_URL}/invitation/example-token`,
   subject: t.subject,
   preview: t.preview.replace("{inviterName}", previewInviterName),
   intro: t.intro.replace("{inviterName}", previewInviterName),
   cta: t.cta,
   fallback: t.fallback,
-};
+} satisfies Props;
