+++ components/emails/contact-inquiry.tsx
@@ -1,22 +1,19 @@
 import { EmailField } from "@/components/emails/base/email-field";
-import { EmailLayout } from "@/components/emails/base/email-layout";
-import { PREVIEW_EMAIL_LAYOUT_COPY, type EmailLayoutCopy } from "@/components/emails/base/email-layout-copy";
-import { DEFAULT_LOCALE, type AppLocale } from "@/i18n/locale-registry";
+import { EmailLayout, type EmailLayoutSharedProps } from "@/components/emails/base/email-layout";
+import { PREVIEW_EMAIL_LAYOUT_PROPS } from "@/components/emails/preview-layout-props";
 
-type Props = {
+type Props = EmailLayoutSharedProps & {
   name: string;
   email: string;
   company?: string;
   message: string;
-  locale: AppLocale;
-  layoutCopy: EmailLayoutCopy;
 };
 
-export default function ContactInquiry({ name, email, company, message, locale, layoutCopy }: Props) {
+export default function ContactInquiry({ name, email, company, message, ...layoutProps }: Props) {
   const subject = "New contact inquiry";
 
   return (
-    <EmailLayout layoutCopy={layoutCopy} locale={locale} preview={subject} title={subject}>
+    <EmailLayout {...layoutProps} preview={subject} title={subject}>
       <EmailField label="From">{`${name} (${email})`}</EmailField>
 
       {company ? <EmailField label="Company">{company}</EmailField> : null}
@@ -27,10 +24,9 @@ export default function ContactInquiry({ name, email, company, message, locale, 
 }
 
 ContactInquiry.PreviewProps = {
-  layoutCopy: PREVIEW_EMAIL_LAYOUT_COPY,
-  locale: DEFAULT_LOCALE,
+  ...PREVIEW_EMAIL_LAYOUT_PROPS,
   name: "Jane Doe",
   email: "jane@example.com",
   company: "Acme Inc.",
   message: "Hi, I'd like to learn more about Customermates for my agency.",
-};
+} satisfies Props;
