+++ components/emails/feedback.tsx
@@ -1,20 +1,17 @@
 import { EmailField } from "@/components/emails/base/email-field";
-import { EmailLayout } from "@/components/emails/base/email-layout";
-import { PREVIEW_EMAIL_LAYOUT_COPY, type EmailLayoutCopy } from "@/components/emails/base/email-layout-copy";
-import { DEFAULT_LOCALE, type AppLocale } from "@/i18n/locale-registry";
+import { EmailLayout, type EmailLayoutSharedProps } from "@/components/emails/base/email-layout";
+import { PREVIEW_EMAIL_LAYOUT_PROPS } from "@/components/emails/preview-layout-props";
 
-type Props = {
+type Props = EmailLayoutSharedProps & {
   feedback: string;
   userEmail: string;
   userName: string;
   subject: string;
-  locale: AppLocale;
-  layoutCopy: EmailLayoutCopy;
 };
 
-export default function Feedback({ feedback, userEmail, userName, subject, locale, layoutCopy }: Props) {
+export default function Feedback({ feedback, userEmail, userName, subject, ...layoutProps }: Props) {
   return (
-    <EmailLayout layoutCopy={layoutCopy} locale={locale} preview={subject} title={subject}>
+    <EmailLayout {...layoutProps} preview={subject} title={subject}>
       <EmailField label="From">{`${userName} (${userEmail})`}</EmailField>
 
       <EmailField label="Type">{subject}</EmailField>
@@ -25,10 +22,9 @@ export default function Feedback({ feedback, userEmail, userName, subject, local
 }
 
 Feedback.PreviewProps = {
-  layoutCopy: PREVIEW_EMAIL_LAYOUT_COPY,
-  locale: DEFAULT_LOCALE,
+  ...PREVIEW_EMAIL_LAYOUT_PROPS,
   feedback: "The CRM is working great! I found it very helpful for managing my contacts.",
   userEmail: "user@example.com",
   userName: "John Doe",
   subject: "Product Feedback",
-};
+} satisfies Props;
