+++ components/emails/new-user-notification.tsx
@@ -1,21 +1,18 @@
 import { EmailField } from "@/components/emails/base/email-field";
-import { EmailLayout } from "@/components/emails/base/email-layout";
-import { PREVIEW_EMAIL_LAYOUT_COPY, type EmailLayoutCopy } from "@/components/emails/base/email-layout-copy";
-import { DEFAULT_LOCALE, type AppLocale } from "@/i18n/locale-registry";
+import { EmailLayout, type EmailLayoutSharedProps } from "@/components/emails/base/email-layout";
+import { PREVIEW_EMAIL_LAYOUT_PROPS } from "@/components/emails/preview-layout-props";
 
-type Props = {
+type Props = EmailLayoutSharedProps & {
   email: string;
   name: string;
   provider?: string;
-  locale: AppLocale;
-  layoutCopy: EmailLayoutCopy;
 };
 
-export default function NewUserNotification({ email, name, provider, locale, layoutCopy }: Props) {
+export default function NewUserNotification({ email, name, provider, ...layoutProps }: Props) {
   const subject = "New user registration";
 
   return (
-    <EmailLayout layoutCopy={layoutCopy} locale={locale} preview={subject} title={subject}>
+    <EmailLayout {...layoutProps} preview={subject} title={subject}>
       <EmailField label="User">{`${name} (${email})`}</EmailField>
 
       {provider ? <EmailField label="Provider">{provider}</EmailField> : null}
@@ -24,9 +21,8 @@ export default function NewUserNotification({ email, name, provider, locale, lay
 }
 
 NewUserNotification.PreviewProps = {
-  layoutCopy: PREVIEW_EMAIL_LAYOUT_COPY,
-  locale: DEFAULT_LOCALE,
+  ...PREVIEW_EMAIL_LAYOUT_PROPS,
   email: "user@example.com",
   name: "John Doe",
   provider: "Google",
-};
+} satisfies Props;
