+++ components/emails/trial-inactivation-reminder.tsx
@@ -1,18 +1,15 @@
 import enMessages from "@/i18n/locales/en.json";
 import { EmailButton } from "@/components/emails/base/email-button";
-import { EmailLayout } from "@/components/emails/base/email-layout";
-import { PREVIEW_EMAIL_LAYOUT_COPY, type EmailLayoutCopy } from "@/components/emails/base/email-layout-copy";
+import { EmailLayout, type EmailLayoutSharedProps } from "@/components/emails/base/email-layout";
 import { EmailLink } from "@/components/emails/base/email-link";
 import { EmailSection } from "@/components/emails/base/email-section";
 import { EmailText } from "@/components/emails/base/email-text";
-import { DEFAULT_LOCALE, type AppLocale } from "@/i18n/locale-registry";
+import { PREVIEW_EMAIL_LAYOUT_PROPS } from "@/components/emails/preview-layout-props";
 import { env } from "@/env";
 
 const CONTACT_HREF = `${env.BASE_URL}/contact`;
 
-type Props = {
-  locale: AppLocale;
-  layoutCopy: EmailLayoutCopy;
+type Props = EmailLayoutSharedProps & {
   greeting: string;
   body: string;
   cta: string;
@@ -25,8 +22,6 @@ type Props = {
 };
 
 export default function TrialInactivationReminder({
-  locale,
-  layoutCopy,
   greeting,
   body,
   cta,
@@ -36,11 +31,12 @@ export default function TrialInactivationReminder({
   subject,
   title,
   href,
+  ...layoutProps
 }: Props) {
   const resolvedHref = href ?? CONTACT_HREF;
 
   return (
-    <EmailLayout layoutCopy={layoutCopy} locale={locale} preview={subject} title={title}>
+    <EmailLayout {...layoutProps} preview={subject} title={title}>
       <EmailText>{greeting}</EmailText>
 
       <EmailText>{body}</EmailText>
@@ -66,8 +62,7 @@ const t = enMessages.TrialInactivationReminder;
 const previewFirstName = "Sofia";
 
 TrialInactivationReminder.PreviewProps = {
-  layoutCopy: PREVIEW_EMAIL_LAYOUT_COPY,
-  locale: DEFAULT_LOCALE,
+  ...PREVIEW_EMAIL_LAYOUT_PROPS,
   greeting: t.greeting.replace("{firstName}", previewFirstName),
   body: t.body,
   cta: t.cta,
@@ -77,4 +72,4 @@ TrialInactivationReminder.PreviewProps = {
   subject: t.subject,
   title: t.title,
   href: CONTACT_HREF,
-};
+} satisfies Props;
